requirements.txt
aiohttp
appdirs
async-timeout
attrs
beautifulsoup4
certifi
chardet
click
discord.py
idna
multidict
mypy-extensions
pathspec
regex
requests
soupsieve
toml
typing-extensions
urllib3
yarl
python-dotenv
python-colored-print